Update all references to get_ram_size to unsigned long
With the recent change to use unsigned longs over longs for get_ram_size
all references have to be updated. Since the old behavior was broken to
begin with, changes are kept as close to begin with for boards. For
boards where there where no consequences and could be safely converted
from long to unsigned long, those have been changed.

Signed-off-by: Oliver Schinagl <oliver@schinagl.nl>