Remove CONS_INDEX from generic A13_MID entry.
As this entry is promoted as generic tablet bootloader and uart turned
on prevents booting at least on some tablets it should be removed.

Separate A13_MID_UART entry is provided as sample for people who are
willing to take apart their tablet and happen to find uart wired as was
on the original A13_MID board.

Signed-off-by: Michal Suchanek <hramrach@gmail.com>